+++ lib/lib/event_id.h
@@ -8,6 +8,8 @@
 #ifndef LIB_EVENT_ID_H_
 #define LIB_EVENT_ID_H_
 
+#include <cstdint>
+
 /**
  * Represents Event Unique Identifier
  * Class can be modified to use other control and data representation such as
@@ -17,10 +19,10 @@ class EventID
 {
 private:
   /// Unique Event identifier
-  int _id;
+  uint8_t _id;
 
 protected:
-  static int id;
+  static uint8_t id;
 
 public:
   EventID();
@@ -33,7 +35,7 @@ public:
    */
   bool is_equal(EventID * event_id);
 
-  const int get_id() const
+  const uint8_t get_id() const
   {
     return _id;
   }
@@ -42,7 +44,7 @@ public:
    * Returns EventID's main counter
    * @return id
    */
-  int get_control_id()
+  uint8_t get_control_id()
   {
     return EventID::id;
   }
